server/container_save_img/requirements.txt
google-cloud-storage
fastapi
uvicorn[standard]
gunicorn